pyproject.toml
[project]
name = "chatbot_langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.29",
    "langsmith>=0.3.30",
    "python-dotenv>=1.1.0",
]
